Extract LocalPackageContainer into separate file
